Change wire communication strategy
Instead of sending data when we want to reload, we now send data to keep
the connection alive (or to check if it's alive), and close the
connection when we want to create a reload scenario.

After the stream is complete, we remove the connection from the list of
listeners.